✨ feat :  브랜드, 제목 검색, 카테고리 제품 총 수량 api 추가 (#128)
* refactor : 사용하지 않는 내용 삭제

* refactor : 쿼리 최적화

* fix : 주석 삭제 및 레포 수정

* refactory : 카테고리 조회 시 offset 사용 안하므로 속도 증가

* feat : 상품 조회시 썸네일 이미지 추가

* feat : 카테고리 예외 코드 및 테스트코드 추가

* feat : 카테고리 예외 코드 및 테스트코드 추가

* feat : 메인 페이지 api 구현

* feat : 메인 페이지 api 구현

* fix : 버그 수정

* feat : 상품 제목명으로 검색 기능 추가 (무한 스크롤)

* fix : 수정

* refactor : 메인 페이지 조회 시 브랜드 명, 카테고리 전송

* feat : 랭킹 구현 중

* feat : 랭킹 API 구현

* refactor : 주석 삭제 및 리턴 타입 변경

* ✅ test : 리팩토링 후 테스트 코드 변경

* ✅ refactor : cross join 회피 및 option 엔티티 수정

* refactor : product param dto, queryDsl 동적 할당 리팩토링

* refactor : product param dto, queryDsl 동적 할당 리팩토링

* fix : 별점 타입 변경

* feat : 리뷰 등록할 떄 상품에 리뷰 평점 평균 저장

* docs : 주문 추가

* fix : 별점 계산 및 타입 수정

* feat : 카테고리 별 상품 총 갯수 요청 api 구현

* ✅ test : 테스트 코드 추가

* ✅ test : 브랜드 상품 전체 조회, 카테고리 별 상품 갯수, 상품 제목 검색 테스트 코드

* 📝 docs : 브랜드 상품 전체, 카테고리 별 전체 상품 갯수, 상품 제목 검색 문서 추가

* 📝 docs : 브랜드 상품 전체 조회 request-parameters 문서 추가

+++ server/src/main/java/gohome/dailydaily/domain/product/entity/Product.java
@@ -37,7 +37,7 @@ public class Product extends BaseTime {
     private Integer price;
 
     @Column(nullable = false)
-    private Float score;
+    private Integer score;
 
     @Column(nullable = false)
     private Integer sale;
@@ -71,13 +71,11 @@ public class Product extends BaseTime {
         this.orderProducts.addAll(List.of(orderProducts));
     }
 
-
-    public void setScore(Float score) {
+    public void setScore(Integer score) {
         this.score = score;
     }
 
     public void updateSale(int count) {
         this.sale += count;
-
     }
 }
